feat: Refactor server setup by creating separate HTTP and WebSocket server files
